style: CharacterCardContent 수정
변경점
- 레이드 이미지 둥근 모서리로 변경
- 이미지를 LazyVerticalGrid로 구성해서 한눈에 보기 편하게 만듦
